fix(package): lodash from devDep. to dependencies
After installing the module via `npm test` the CLI usage fails because the dependency "lodash" was missing.
The reason was that in the `package.json` "lodash" was configured as dev-dependency and not install via the simple `npm install`.

The fix was simply to move the configuration into the normal dependencies - so it gets installed initially.